refactor: reorganización de componentes siguiendo mejores prácticas
